// src/app/types/comment.d.ts
export interface IComment {
  _id: string;
  content: string;
  course_id: string;
  createdAt: string;
  isReply: boolean;
  lesson_id: string;
  numberDisLike: number;
  numberLike: number;
  rating: number;
  replyIds: IComment[];
  user_id: string;
  user_info: {
    email: string;
    fullName: string;
    avatar: string;
  };
}